Made some changes to make game scalable
The old method was just horrible so i refactored a little to get a new
way to scale the game. It worked really well with my initial tests and
so far no problems with it!
Also added a screen to end screen to show the last frame where you died!
Also lowered the spawn rate of entities a bit.